Mostly changing how Git is executed.
I need to support windows and linux modes and so we'll need better logic
in how we infer the path to Git. For now we fall back on the linux
style when the given path is missing.

Also there was a bug in how things were being run before and so the
Git command is now an array so we can do stuff like /usr/bin/env
/usr/bin/git and such and things will work correctly.

I had another problem with PROJECT_NAME being unset and so there's a
hack in there to fix that for now. I'll address that once I've got the
bigger issues worked out.